move authority utils
