+++ pages/browse/index.tsx
@@ -1,6 +1,6 @@
 import { Flex, List, Text } from "@nypl/design-system-react-components"
 import RCLink from "../../src/components/Links/RCLink/RCLink"
-import { run } from "./sierraUtils"
+import { run } from "../../src/models/Headings/sierraUtils"
 import Heading from "../../src/models/Headings/Heading"
 import type AuthorityVarfield from "../../src/models/Headings/AuthorityVarfield"
